[Refactor] 온보딩 검증 validator 추가
