+++ src/main/java/DNBN/spring/service/ArticleService/ArticleCommandServiceImpl.java
@@ -23,8 +23,8 @@ import DNBN.spring.repository.CategoryRepository.CategoryRepository;
 import DNBN.spring.repository.MemberRepository.MemberRepository;
 import DNBN.spring.repository.PlaceRepository.PlaceRepository;
 import DNBN.spring.repository.RegionRepository.RegionRepository;
-import DNBN.spring.validation.ContentLengthValidator;
-import DNBN.spring.validation.TitleLengthValidator;
+import DNBN.spring.validation.validator.ContentLengthValidator;
+import DNBN.spring.validation.validator.TitleLengthValidator;
 import DNBN.spring.web.dto.ArticleRequestDTO;
 import DNBN.spring.web.dto.ArticleUpdateRequestDTO;
 import DNBN.spring.web.dto.ArticleWithLocationRequestDTO;
