+++ src/main/java/DNBN/spring/service/CommentService/CommentCommandServiceImpl.java
@@ -13,8 +13,7 @@ import DNBN.spring.repository.ArticleRepository.ArticleRepository;
 import DNBN.spring.repository.CommentRepository.CommentRepository;
 import DNBN.spring.repository.MemberRepository.MemberRepository;
 import DNBN.spring.repository.NotificationRepository.NotificationRepository;
-import DNBN.spring.validation.ContentLengthValidator;
-import DNBN.spring.validation.TitleLengthValidator;
+import DNBN.spring.validation.validator.ContentLengthValidator;
 import DNBN.spring.web.dto.CommentRequestDTO;
 import DNBN.spring.web.dto.CommentResponseDTO;
 import DNBN.spring.web.dto.CommentUpdateRequestDTO;
